Rm broken Data link, rm Research menu, put Sharing in About menu
The data link used to be PersonalGenomes.org but now goes directly to
PGP Harvard's data (for now at least). So I removed the Data link in
the menubar (because I think that should be PG.org level stuff). That
left the Research menu with a single item (Sharing Policies) so I moved
that item to the About menu and removed the Research menu.