Add a translation icon & custom icon font
To make a custom font from SVGs: save new "letter" in Inkscape
using the Fontforge template. Then open Fontforge, make a new
font, double click on a letter and import the SVG to set it as
the path for that letter.

Use "Generate Font" to save as ttf, otf, woff, and svg fonts.
Then used mkeot command from eot-utils to make eot font.