Add language-change menu
Language is changed via POST... which as I understand it, means
it needs to use a <form>. That makes it hard to match the style
of Bootstraps dropdown lists.

Because it's complicated to try to match the Bootstrap dropdown
list links style for the form submit button, I've let it be a
small button stating "Change to X".